# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "description": "[RFC 7807](https://www.rfc-editor.org/rfc/rfc7807) problem document (`application/problem+json`).\n\nCanonical error body for HTTP 4xx/5xx responses. Extension members include `code`.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (extension member).",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable explanation.",
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "description": "Optional URI reference that identifies the specific occurrence.",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "description": "HTTP status code.",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "description": "Short, stable summary of the problem class.",
            "type": "string"
          },
          "type": {
            "description": "URI reference identifying the problem type.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SessionBody": {
        "description": "Wire representation of a session. `user` is a [`TeamUser`] link unless the client\npasses `expand=user`, in which case it is the full [`User`] object.",
        "example": {
          "created_at": "2024-01-01T12:00:00Z",
          "expires_at": "2025-01-01T12:00:00Z",
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "user": {
            "email": "singer@example.com",
            "id": "user-1"
          }
        },
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserBody"
          }
        },
        "required": [
          "id",
          "user",
          "created_at",
          "expires_at"
        ],
        "type": "object"
      },
      "SessionUserBody": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/User",
            "description": "Try full user first so link-shaped JSON (missing required user fields) falls through to [`TeamUser`]."
          },
          {
            "$ref": "#/components/schemas/TeamUser"
          }
        ]
      },
...
      "TeamUser": {
        "description": "User slice returned on team **GET** (`id` + `email` only, same naming as `User`).",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Session cookie returned after a successful authentication flow",
        "in": "cookie",
        "name": "sso_session",
        "type": "apiKey"
      },
      "SessionToken": {
        "description": "Session override using `Authorization: Bearer <session>` header",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Worship Viewer API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/users/me/session": {
      "get": {
        "operationId": "get_current_session_for_user",
        "parameters": [
          {
            "description": "Optional `user` to embed full user (default: `id`+`email` link).",
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBody"
                }
              }
            },
            "description": "Returns the session credential used for this request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Session not found"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch session"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/users/me/sessions": {
      "get": {
        "operationId": "get_sessions_for_current_user",
        "parameters": [
          {
            "description": "Zero-based page; defaults to 0. `X-Total-Count` is the total before paging (`list-pagination.md`).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Items per page. Must be 1–500. Defaults to 50 when omitted.",
            "example": 50,
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "Optional case-insensitive substring on session id, user id, or user email. Whitespace-only is treated as absent.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional comma-separated relations (`user` = embed full user on each session; default is `id`+`email` link only).",
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionBody"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Returns active sessions for the current user. `X-Total-Count` is the total before paging."
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid pagination parameters"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to list sessions for current user"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/sessions/{id}": {
      "delete": {
        "operationId": "delete_session_for_current_user",
        "parameters": [
          {
            "description": "Session identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Session not found for current user"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to delete session"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "get_session_for_current_user",
        "parameters": [
          {
            "description": "Session identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Optional `user` to embed full user (default: `id`+`email` link).",
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBody"
                }
              }
            },
            "description": "Returns a session for the current user"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Session not found for current user"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "API rate limit exceeded; see `Retry-After` and `X-RateLimit-*` response headers"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Failed to fetch session"
          }
        },
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SessionToken": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same origin as the web app (override per deployment).",
      "url": "/"
    },
    {
      "description": "Production deployment (public web app).",
      "url": "https://app.worshipviewer.com"
    }
  ],
...
}